Resizable: Move resizable into widgets folder
